[HOTSPOT-288] fix: userStore id에서 subId로 수정
