+++ apps/web/src/entities/policy/ui/PolicyUserCard.tsx
@@ -16,7 +16,7 @@ interface PolicyUserCardProps {
 export const PolicyUserCard = ({ familyId, data }: PolicyUserCardProps) => {
   const [isAccordianOpen, setIsAccordianOpen] = useState(false);
   const { open } = useModal();
-  const isMe = data.memberId === useUserStore().id;
+  const isMe = data.memberId === useUserStore().subId;
 
   const handleOpenModal = useCallback(() => {
     open('policyDetailModal', {
