+++ apps/web/src/features/policy-order/ui/OrderItem.tsx
@@ -15,7 +15,7 @@ interface OrderItemProps {
 }
 
 export const OrderItem = ({ member, index, isEditing, onMove, isLast }: OrderItemProps) => {
-  const myId = useUserStore().id;
+  const myId = useUserStore().subId;
 
   return (
     <Draggable draggableId={`item${member.subId}`} index={index} isDragDisabled={!isEditing}>
